Convert to plugin2.
Next Dancer2 release includes a major overhaul to Plugin (known as
plugin2) which breaks this plugin. Attempts have been made to prevent
breakage of existing plugins but this is not possible in all cases (like
this one) so a new release of this plugin will be needed as soon as next
Dancer2 release occurs.